Update shinytest.R
Issue with shinytest. 
shinytest::testApp("./") "1 dot" causes the shinytest workflow to fail, but the deploy app workflow is completed successfully.
shinytest::testApp("../") "2 dots" causes the shinytest workflow to complete successfully, but then the deploy app workflow fails.
ShinyTest Errors (2 Errors) seem to be related to MacOS, However tested the app on an IPhone and seems to be working correctly (and has been updated) with shinytest::testApp("./").

So proceeding with shinytest::testApp("./") even though this causes the shinytest workflow to fail with 2 errors.